+++ base/gsparaml.c
@@ -294,7 +294,7 @@ process_array(gs_memory_t *mem, gs_c_param_list *plist, gs_param_name key, char 
                     if (new_max == 0)
                         new_max = 32;
                     if (array_data == NULL) {
-                        array_data = (char *)gs_alloc_bytes(mem, sizeof(gs_param_string) * new_max, "param string array in param parsing");
+                        array_data = (char *)gs_alloc_bytes(mem, sizeof(gs_param_string) * (size_t)new_max, "param string array in param parsing");
                         if (array_data == NULL) {
                             code = gs_error_VMerror;
                             break;
@@ -357,7 +357,7 @@ process_array(gs_memory_t *mem, gs_c_param_list *plist, gs_param_name key, char 
                     if (new_max == 0)
                         new_max = 32;
                     if (array_data == NULL) {
-                        array_data = (char *)gs_alloc_bytes(mem, sizeof(gs_param_string) * new_max, "param string array in param parsing");
+                        array_data = (char *)gs_alloc_bytes(mem, sizeof(gs_param_string) * (size_t)new_max, "param string array in param parsing");
                         if (array_data == NULL) {
                             code = gs_error_VMerror;
                             break;
@@ -393,7 +393,7 @@ process_array(gs_memory_t *mem, gs_c_param_list *plist, gs_param_name key, char 
                     if (new_max == 0)
                         new_max = 32;
                     if (array_data == NULL) {
-                        array_data = (char *)gs_alloc_bytes(mem, sizeof(gs_param_string) * new_max, "param string array in param parsing");
+                        array_data = (char *)gs_alloc_bytes(mem, sizeof(gs_param_string) * (size_t)new_max, "param string array in param parsing");
                         if (array_data == NULL) {
                             code = gs_error_VMerror;
                             break;
@@ -451,7 +451,7 @@ process_array(gs_memory_t *mem, gs_c_param_list *plist, gs_param_name key, char 
 
                     if (array_type == gs_param_type_int_array && !integer) {
                         ints = (int *)array_data;
-                        floats = (float *)gs_alloc_bytes(mem, sizeof(float) * array_max, "param string array in param parsing");
+                        floats = (float *)gs_alloc_bytes(mem, sizeof(float) * (size_t)array_max, "param string array in param parsing");
                         if (floats == NULL){
                             code = gs_error_VMerror;
                             break;
@@ -471,13 +471,13 @@ process_array(gs_memory_t *mem, gs_c_param_list *plist, gs_param_name key, char 
                             array_type = integer ? gs_param_type_int_array : gs_param_type_float_array;
                         }
                         if (array_data == NULL) {
-                            array_data = (char *)gs_alloc_bytes(mem, sizeof(size_me) * new_max, "param string array in param parsing");
+                            array_data = (char *)gs_alloc_bytes(mem, sizeof(size_me) * (size_t)new_max, "param string array in param parsing");
                             if (array_data == NULL) {
                                 code = gs_error_VMerror;
                                 break;
                             }
                         } else {
-                            char *new_array = (char *)gs_resize_object(mem, array_data, sizeof(size_me) * new_max, "param string array in param parsing");
+                            char *new_array = (char *)gs_resize_object(mem, array_data, sizeof(size_me) * (size_t)new_max, "param string array in param parsing");
                             if (new_array == NULL) {
                                 code = gs_error_VMerror;
                                 break;
